requirements.txt
numpy~=2.4.3
opencv-python~=4.13.0.92
fastapi~=0.135.1
pillow~=12.1.1
torch~=2.10.0
transformers~=5.3.0
albumentations~=2.0.8
segmentation_models_pytorch~=0.5.0
torchvision~=0.25.0